@@ -2,27 +2,33 @@
   "name": "angular-samples",
   "scripts": {
     "build-all": "bazel build //src/apps/api && bazel build //src/apps/backend",
-    "start": "concurrently \"ibazel run //src/apps/api\" \"ibazel run //src/apps/backend\"",
-    "postinstall": "ngc -p angular-metadata.tsconfig.json"
+    "run-backend": "ibazel run //src/apps/backend/src:src",
+    "start": "concurrently \"ibazel run //src/apps/api\" \"npm run run-backend\"",
+    "postinstall": "ngcc"
   },
   "devDependencies": {
-    "@angular/bazel": "8.0.0",
-    "@angular/cli": "8.0.0",
-    "@angular/compiler": "8.0.0",
-    "@angular/compiler-cli": "8.0.0",
+    "@angular/bazel": "9.0.0",
+    "@angular/compiler": "9.0.0",
+    "@angular/compiler-cli": "9.0.0",
+    "@babel/cli": "^7.6.0",
+    "@babel/core": "^7.6.0",
     "@bazel/buildifier": "0.22.0",
-    "@bazel/ibazel": "0.10.1",
-    "@bazel/typescript": "0.27.12",
+    "@bazel/ibazel": "0.11.1",
+    "@bazel/typescript": "1.2.0",
+    "@bazel/terser": "^1.3.0",
     "@bazel/karma": "0.27.12",
-    "@bazel/bazel": "0.25.1",
+    "@bazel/bazel": "2.0.0",
+    "@babel/preset-env": "^7.6.0",
+    "@bazel/rollup": "^1.3.0",
     "@types/node": "^10.12.0",
     "@types/react": "16.4.1",
     "compression": "1.7.2",
     "concurrently": "4.1.0",
-    "reload": "2.3.1",
-    "typescript": "3.4.5",
-    "webpack": "4.20.2",
-    "webpack-cli": "3.1.2"
+    "typescript": "3.6.4",
+    "rollup": "^1.21.4",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-node-resolve": "^5.2.0",
+    "terser": "4.3.1"
   },
   "dependencies": {
     "tslib": "1.9.3",
@@ -30,18 +36,17 @@
     "express": "4.16.3",
     "jquery": "3.4.0",
     "jquery-ui-dist": "1.12.1",
-    "react": "16.4.2",
-    "react-dom": "16.4.2",
     "request": "2.87.0",
     "zone.js": "0.9.0",
-    "@angular/common": "8.0.0",
-    "@angular/core": "8.0.0",
-    "@angular/forms": "8.0.0",
-    "@angular/platform-browser": "8.0.0",
-    "@angular/platform-browser-dynamic": "8.0.0",
-    "@angular/router": "8.0.0",
-    "@angular/upgrade": "8.0.0",
-    "@angular/animations": "8.0.0",
+    "@angular/common": "9.0.0",
+    "@angular/core": "9.0.0",
+    "@angular/forms": "9.0.0",
+    "@angular/platform-browser": "9.0.0",
+    "@angular/platform-browser-dynamic": "9.0.0",
+    "@angular/router": "9.0.0",
+    "@angular/upgrade": "9.0.0",
+    "@angular/animations": "9.0.0",
+    "@bazel/rollup": "^1.3.0",
     "rxjs": "6.4.0",
     "systemjs": "0.21.6"
   }
