@@ -1,10 +1,4 @@
-export {
-  IFriend,
-  IFriends,
-} from 'angular_samples/src/apps/models/friend/friend_ts_proto';
-export {ILog} from 'angular_samples/src/apps/models/log/log_ts_proto';
-export {
-  IPerson,
-  IPersons,
-} from 'angular_samples/src/apps/models/person/person_ts_proto';
-export {ICar} from 'angular_samples/src/apps/models/car/car_ts_proto';
+export {IFriend} from './friend/friend';
+export {ILog} from './log/log';
+export {IPerson} from './person/person';
+export {ICar} from './car/car';
