@@ -1,7 +1,5 @@
 export class FriendsService {
   getFriends() {
-    return fetch('/api/friends')
-           .then(data => data.json())
-           .then(data => data.friends);
+    return fetch('/api/friends').then(data => data.json());
   }
-} 
+}
